+++ client/src/App.tsx
@@ -1,4 +1,3 @@
-// import React from "react";
 import MainRoutes from "./routes/MainRoutes";
 
 function App() {
