+++ client/src/pages/MainPage/List/List.css.ts
@@ -60,39 +60,11 @@ export const weatherIcon = style({
   fontSize: vars.fontSizing.H4,
 });
 
-export const trashIcon = style({
-  width: "10%",
-  textAlign: "center",
-  cursor: "pointer",
-  marginLeft: "2.5%",
-  fontSize: vars.fontSizing.H7,
-});
-
 export const taskListArea = style({
   height: "52vh",
   marginRight: "2%",
   overflowY: "scroll",
   ...myScrollbar,
-});
-
-export const task = style({
-  height: "15%",
-  alignItems: "center",
-  display: "flex",
-  marginTop: "1%",
-});
-
-export const taskDetail = style({
-  width: "80%",
-  height: "60%",
-  minWidth: "50px",
-  minHeight: "40px",
-  textAlign: "center",
-  alignItems: "center",
-  backgroundColor: vars.color.brightText,
-  marginLeft: "5%",
-  cursor: "pointer",
-  borderRadius: vars.radius.small,
 });
 
 export const newTaskButton = style({
