+++ client/src/redux/index.ts
@@ -1,0 +1,10 @@
+import { configureStore } from "@reduxjs/toolkit";
+import reducer from "./reducer/reducer";
+
+const store = configureStore({
+    reducer
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
+export default store;
